Final draft - moved component and styling matches
