--- packages/components/src/embed-container/gallery-style.scss
@@ -1,159 +1,0 @@
-// Carousel overlay that covers the entire screen
-.reader-image-carousel-overlay {
-	position: fixed;
-	top: 0;
-	left: 0;
-	width: 100%;
-	height: 100%;
-	background-color: #000;
-	z-index: 9999;
-	display: flex;
-	align-items: flex-start;
-	justify-content: center;
-}
-
-// Main container for the carousel content
-.reader-image-carousel-container {
-	position: relative;
-	width: 100%;
-	height: 100%;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	margin: 0 50px;
-
-	// Remove margins on mobile
-	@media (max-width: 768px) {
-		margin: 0;
-	}
-}
-
-// Swiper wrapper styles
-.reader-image-carousel-wrap {
-	width: 100%;
-	height: calc(100% - 64px);
-	position: relative;
-
-	.swiper-slide {
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		
-		img {
-			max-width: 100%;
-			max-height: 100%;
-			object-fit: contain;
-		}
-	}
-
-	// Navigation arrows
-	.swiper-button-next,
-	.swiper-button-prev {
-		color: #FFF;
-		background-color: rgba(0, 0, 0, 0.5);
-		border-radius: 4px;
-		width: 30px;
-		height: 30px;
-		top: 50%;
-
-		&::after {
-			font-size: 12px;
-			font-weight: bold;
-		}
-
-		&:hover {
-			background-color: #999;
-		}
-
-		// Hide on mobile for touch navigation
-		@media (max-width: 768px) {
-			display: none;
-		}
-	}
-
-	.swiper-button-next {
-		right: 40px;
-	}
-
-	.swiper-button-prev {
-		left: 40px;
-	}
-}
-
-// Close button in top right
-.reader-image-carousel-close {
-	position: absolute;
-	top: 20px;
-	right: 20px;
-	background: rgba(0, 0, 0, 0.5);
-	border: none;
-	border-radius: 4px;
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	z-index: 10001;
-
-	&:hover {
-		background-color: #999;
-	}
-}
-
-// Footer with pagination - full width across bottom
-.reader-image-carousel-footer {
-	position: absolute;
-	bottom: 0;
-	left: 0;
-	right: 0;
-	width: 100%;
-	height: 64px;
-	color: #FFF;
-	z-index: 10001;
-	padding: 16px;
-}
-
-.reader-image-carousel-pagination-container {
-	display: flex;
-	align-items: center;
-	justify-content: flex-start;
-}
-
-.reader-image-carousel-pagination {
-	background: rgba(0, 0, 0, 0.5);
-	padding: 8px 16px;
-	border-radius: 16px;
-	font-size: 14px;
-	font-weight: 500;
-
-	span {
-		color: #FFF;
-	}
-}
-
-// Ensure all text and icons are white
-.reader-image-carousel-container * {
-	color: #FFF;
-}
-
-// Additional mobile responsiveness
-@media (max-width: 768px) {
-	.jp-carousel-overlay {
-		padding-top: 20px;
-	}
-
-	.reader-image-carousel-wrap {
-		height: calc(100vh - 100px);
-	}
-
-	.reader-image-carousel-close {
-		top: 10px;
-		right: 10px;
-		width: 40px;
-		height: 40px;
-	}
-
-	.reader-image-carousel-footer {
-		bottom: 8px;
-		left: 8px;
-	}
-}
