+++ packages/components/src/image-carousel/style.scss
@@ -1,0 +1,123 @@
+:root {
+	--swiper-navigation-sides-offset: 40px;
+}
+
+.reader-image-carousel-overlay {
+	position: fixed;
+	inset: 0;
+	z-index: 9999;
+	display: flex;
+	align-items: flex-start;
+	justify-content: center;
+	background-color: #000;
+
+	@media (max-width: 768px) {
+		padding-top: 20px;
+	}
+}
+
+.reader-image-carousel-container {
+	width: 100%;
+	height: 100%;
+	align-items: start;
+}
+
+.reader-image-carousel-wrap {
+	position: relative;
+	width: 100%;
+	height: calc(100% - 64px);
+
+	@media (max-width: 768px) {
+		height: calc(100vh - 100px);
+	}
+
+	.swiper-slide {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		
+		img {
+			max-width: 100%;
+			max-height: 100%;
+			object-fit: contain;
+		}
+	}
+
+	.swiper-button-next,
+	.swiper-button-prev {
+		top: calc(50% + 20px);
+		width: 30px;
+		height: 30px;
+		color: #999;
+		background-color: rgba(0, 0, 0, 0.5);
+		border-radius: 4px;
+
+		&::after {
+			font-size: 12px;
+			font-weight: bold;
+		}
+
+		&:hover {
+			color: #fff;
+			background-color: #999;
+		}
+
+		@media (max-width: 768px) {
+			display: none;
+		}
+	}
+}
+
+.reader-image-carousel-close {
+	position: absolute;
+	top: 20px;
+	right: 30px;
+	z-index: 10001;
+	border: none;
+	background: rgba(0, 0, 0, 0.5);
+	border-radius: 4px;
+	cursor: pointer;
+
+	&:hover {
+		background-color: #999;
+	}
+
+	@media (max-width: 768px) {
+		top: 10px;
+		right: 10px;
+		width: 40px;
+		height: 40px;
+	}
+}
+
+.reader-image-carousel-footer {
+	position: absolute;
+	bottom: 0;
+	left: 0;
+	right: 0;
+	z-index: 10001;
+	display: flex;
+	align-items: center;
+	width: 100%;
+	height: 64px;
+	padding: 16px;
+	color: #999;
+	font-size: 14px;
+	font-weight: 500;
+	box-sizing: border-box;
+
+	> * {
+		width: 33.333%;
+		padding: 8px 12px;
+		box-sizing: border-box;
+	}
+
+	.reader-image-carousel-caption {
+		text-align: center;
+	}
+
+	@media (max-width: 768px) {
+		bottom: 8px;
+		left: 8px;
+	}
+}
